pyproject.toml
[tool.poetry]
name = "telco-dwh-project"
version = "0.1.0"
description = ""
authors = ["Erikas Leonenka <e.leonenka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
